LPS-199593 inline MembershipsUserFieldExpressionHandler into Registrator class, removing service for serviceReference and passing properties by hashMapDictionaryBuilder on _serviceRegistration, add constructor in handler class passing the processIndex atribute, modify tests
